loading spinner for strategy pages
