+++ app/[projectName]/[strategyName]/[deployment]/loading.tsx
@@ -1,0 +1,6 @@
+import { Spinner } from "flowbite-react";
+
+export default function Loading() {
+  // You can add any UI inside Loading, including a Skeleton.
+  return <Spinner />;
+}
