+++ app/_components/WebappFrame.tsx
@@ -72,27 +72,30 @@ const WebappFrame = ({ dotrainText, deploymentOption }: props) => {
   const [currentState, setCurrentState] = useState<FrameState>(
     urlState || defaultState
   );
-  const [fetchingTokens, setFetchingTokens] = useState(true);
+  const [fetchingTokens, setFetchingTokens] = useState(false);
   const [error, setError] = useState<string | null>(null);
   const [inputText, setInputText] = useState<string>("");
 
   useEffect(() => {
-    const fetchTokenInfos = async () => {
-      try {
-        const tokenInfos = await getTokenInfos(yamlData);
-        setCurrentState((prevState) => ({
-          ...prevState,
-          tokenInfos,
-        }));
-      } catch (e) {
-        setError("Failed to fetch token information");
-      } finally {
-        setFetchingTokens(false);
-      }
-    };
-
-    fetchTokenInfos();
-  }, [yamlData]);
+    if (!fetchingTokens && currentState.tokenInfos.length === 0) {
+      const fetchTokenInfos = async () => {
+        setFetchingTokens(true);
+        try {
+          const tokenInfos = await getTokenInfos(yamlData);
+          setCurrentState((prevState) => ({
+            ...prevState,
+            tokenInfos,
+          }));
+        } catch (e) {
+          console.error(e);
+          setError("Failed to fetch token information");
+        } finally {
+          setFetchingTokens(false);
+        }
+      };
+      fetchTokenInfos();
+    }
+  }, [yamlData, fetchingTokens, currentState.tokenInfos.length]);
 
   const handleButtonClick = async (buttonData: any) => {
     setError(null);
@@ -136,75 +139,77 @@ const WebappFrame = ({ dotrainText, deploymentOption }: props) => {
   const buttonsData = generateButtonsData(yamlData, currentState);
 
   return (
-    <div className="flex-grow flex-col flex w-full ">
-      <div className="w-full top-0">
-        <ProgressBar currentState={currentState} />
-      </div>
-      <FrameImage currentState={currentState} />
-      {currentState.textInputLabel && (
-        <div className="flex justify-center mb-4">
-          <input
-            className="border-gray-200 rounded-lg border p-2 w-full max-w-96"
-            type="number"
-            placeholder={currentState.textInputLabel}
-            value={inputText}
-            onChange={(e) => setInputText(e.target.value)}
-          />
-          <br />
+    !fetchingTokens && (
+      <div className="flex-grow flex-col flex w-full ">
+        <div className="w-full top-0">
+          <ProgressBar currentState={currentState} />
         </div>
-      )}
-      <div className="flex flex-wrap gap-2 justify-center md:pb-20 pb-8 px-8 pt-10">
-        {buttonsData.map((buttonData) => {
-          return buttonData.buttonValue === "finalSubmit" ? (
-            <div className="flex gap-2 flex-wrap justify-center">
-              <SubmissionModal
+        <FrameImage currentState={currentState} />
+        {currentState.textInputLabel && (
+          <div className="flex justify-center mb-4">
+            <input
+              className="border-gray-200 rounded-lg border p-2 w-full max-w-96"
+              type="number"
+              placeholder={currentState.textInputLabel}
+              value={inputText}
+              onChange={(e) => setInputText(e.target.value)}
+            />
+            <br />
+          </div>
+        )}
+        <div className="flex flex-wrap gap-2 justify-center md:pb-20 pb-8 px-8 pt-10">
+          {buttonsData.map((buttonData) => {
+            return buttonData.buttonValue === "finalSubmit" ? (
+              <div className="flex gap-2 flex-wrap justify-center">
+                <SubmissionModal
+                  key={buttonData.buttonText}
+                  buttonText={buttonData.buttonText}
+                  yamlData={yamlData}
+                  currentState={currentState}
+                  dotrainText={dotrainText}
+                  setError={setError}
+                />
+                <ShareStateAsUrl currentState={currentState} />
+              </div>
+            ) : (
+              <Button
+                color="primary"
+                size="sm"
                 key={buttonData.buttonText}
-                buttonText={buttonData.buttonText}
-                yamlData={yamlData}
-                currentState={currentState}
-                dotrainText={dotrainText}
-                setError={setError}
-              />
-              <ShareStateAsUrl currentState={currentState} />
-            </div>
-          ) : (
-            <Button
-              color="primary"
-              size="sm"
-              key={buttonData.buttonText}
-              onClick={async () => {
-                await handleButtonClick(buttonData);
-              }}
-            >
-              {buttonData.buttonText}
-            </Button>
-          );
-        })}
-        <div className="flex w-full justify-center">
-          {currentState.error ? (
-            <div style={{ display: "flex", justifyContent: "center" }}>
-              <span className="text-red-600">{currentState.error}</span>
-            </div>
-          ) : (
-            ""
-          )}
+                onClick={async () => {
+                  await handleButtonClick(buttonData);
+                }}
+              >
+                {buttonData.buttonText}
+              </Button>
+            );
+          })}
+          <div className="flex w-full justify-center">
+            {currentState.error ? (
+              <div style={{ display: "flex", justifyContent: "center" }}>
+                <span className="text-red-600">{currentState.error}</span>
+              </div>
+            ) : (
+              ""
+            )}
+          </div>
         </div>
+        <Dialog open={!!error}>
+          <DialogContent className="bg-white flex flex-col items-center">
+            <TriangleAlert color="red" />
+            <div className="w-full text-center">{error}</div>
+            <DialogClose asChild>
+              <button
+                className="bg-blue-500 hover:bg-blue-700 text-white py-2 px-4 rounded-xl transition-colors"
+                onClick={() => setError(null)}
+              >
+                Close
+              </button>
+            </DialogClose>
+          </DialogContent>
+        </Dialog>
       </div>
-      <Dialog open={!!error}>
-        <DialogContent className="bg-white flex flex-col items-center">
-          <TriangleAlert color="red" />
-          <div className="w-full text-center">{error}</div>
-          <DialogClose asChild>
-            <button
-              className="bg-blue-500 hover:bg-blue-700 text-white py-2 px-4 rounded-xl transition-colors"
-              onClick={() => setError(null)}
-            >
-              Close
-            </button>
-          </DialogClose>
-        </DialogContent>
-      </Dialog>
-    </div>
+    )
   );
 };
